Move upstream/utility.h/cc to http/filter_utility.h/cc.
