+++ source/common/grpc/grpc_web_filter.cc
@@ -5,8 +5,8 @@
 #include "common/common/base64.h"
 #include "common/common/utility.h"
 #include "common/grpc/common.h"
+#include "common/http/filter_utility.h"
 #include "common/http/headers.h"
-#include "common/upstream/utility.h"
 
 #include "spdlog/spdlog.h"
 
@@ -167,7 +167,7 @@ Http::FilterTrailersStatus GrpcWebFilter::encodeTrailers(Http::HeaderMap& traile
 }
 
 void GrpcWebFilter::setupStatTracking(const Http::HeaderMap& headers) {
-  cluster_ = Upstream::Utility::resolveClusterInfo(decoder_callbacks_, cm_);
+  cluster_ = Http::FilterUtility::resolveClusterInfo(decoder_callbacks_, cm_);
   if (!cluster_) {
     return;
   }
